+++ models/invoice_coupon.js
@@ -1,17 +1,21 @@
-'use strict';
+"use strict";
 
 // Import Config
-const constants = require('../config/constants');
+const constants = require("../config/constants");
 
 module.exports = (sequelize, DataTypes) => {
-  var invoiceCoupon = sequelize.define('invoice_coupon', {
-    invoice_no: DataTypes.INTEGER,
-    coupon_code: DataTypes.BIGINT,
-    applicable_on: DataTypes.STRING,
-    discount: DataTypes.FLOAT,
-    cashback: DataTypes.FLOAT,
-    status: DataTypes.BOOLEAN,
-  }, {});
+  var invoiceCoupon = sequelize.define(
+    "invoice_coupon",
+    {
+      invoice_no: DataTypes.INTEGER,
+      coupon_code: DataTypes.BIGINT,
+      applicable_on: DataTypes.STRING,
+      discount: DataTypes.DECIMAL,
+      cashback: DataTypes.DECIMAL,
+      status: DataTypes.BOOLEAN
+    },
+    {}
+  );
   invoiceCoupon.associate = function(models) {
     // associations can be defined here
   };
