+++ models/manual_discount.js
@@ -1,16 +1,20 @@
-'use strict';
+"use strict";
 
 // Import Config
-const constants = require('../config/constants');
+const constants = require("../config/constants");
 
 module.exports = (sequelize, DataTypes) => {
-  var manualDiscount = sequelize.define('manual_discount', {
-    store_id: DataTypes.INTEGER,
-    warehouse_user_id: DataTypes.INTEGER,
-    invoice_no: DataTypes.INTEGER,
-    discount_amount: DataTypes.FLOAT,
-    status: DataTypes.BOOLEAN
-  }, {});
+  var manualDiscount = sequelize.define(
+    "manual_discount",
+    {
+      store_id: DataTypes.INTEGER,
+      warehouse_user_id: DataTypes.INTEGER,
+      invoice_no: DataTypes.INTEGER,
+      discount_amount: DataTypes.DECIMAL,
+      status: DataTypes.BOOLEAN
+    },
+    {}
+  );
   manualDiscount.associate = function(models) {
     // associations can be defined here
   };
